Styling Changes, Moved SearchBar to activityFinding
