+++ client/src/components/activityFinding/SearchBar.jsx
